Add file upload using multer and host static files

+++ controllers/admin.js
@@ -1,4 +1,5 @@
 const { validationResult } = require('express-validator');
+const fileHelper = require('../util/file');
 const Product = require('../models/product');
 
 exports.getAddProductPage = (req, res, next) => {
@@ -14,10 +15,26 @@ exports.getAddProductPage = (req, res, next) => {
 
 exports.postAddProduct = (req, res, next) => {
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
-  
+
+  if (!image) {
+    return res.status(422).render('admin/edit-product', {
+      pageTitle: 'Add Product',
+      path: '/admin/add-product',
+      editing: false,
+      hasError: true,
+      validationErrors: [],
+      errorMessage: 'Attached file is not an image',
+      product: {
+        title: title,
+        price: price,
+        description: description,
+      },
+    });
+  }
+
   const errors = validationResult(req);
   if (!errors.isEmpty()) {
     return res.status(422).render('admin/edit-product', {
@@ -30,11 +47,13 @@ exports.postAddProduct = (req, res, next) => {
       product: {
         title: title,
         price: price,
-        imageUrl: imageUrl,
         description: description,
-      }
+      },
     });
   }
+
+  const imageUrl = image.path;
+
   const product = new Product({
     title: title,
     price: price,
@@ -75,7 +94,7 @@ exports.getEditProduct = (req, res, next) => {
         errorMessage: '',
       });
     })
-    .catch((err) =>{
+    .catch((err) => {
       const error = new Error(err);
       error.httpStatus = 500;
       return next(error);
@@ -85,7 +104,7 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const prodId = req.body.productId;
   const title = req.body.title;
-  const imageUrl = req.body.imageUrl;
+  const image = req.file;
   const price = req.body.price;
   const description = req.body.description;
 
@@ -101,10 +120,9 @@ exports.postEditProduct = (req, res, next) => {
       product: {
         title: title,
         price: price,
-        imageUrl: imageUrl,
         description: description,
         _id: prodId,
-      }
+      },
     });
   }
 
@@ -115,7 +133,10 @@ exports.postEditProduct = (req, res, next) => {
       }
       product.title = title;
       product.price = price;
-      product.imageUrl = imageUrl;
+      if (image) {
+        fileHelper.deleteFile(product.imageUrl);
+        product.imageUrl = image.path;
+      }
       product.description = description;
       return product.save().then((result) => {
         console.log('UPDATED PRODUCT!!!');
@@ -130,7 +151,7 @@ exports.postEditProduct = (req, res, next) => {
 };
 
 exports.getProducts = (req, res, next) => {
-  Product.find({ userId: req.user._id})
+  Product.find({ userId: req.user._id })
     .populate('userId')
     .then((products) => {
       res.render('admin/products', {
@@ -148,7 +169,15 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const productId = req.body.productId;
-  Product.deleteOne({_id: productId, userId: req.user._id})
+  Product.findById(productId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error('Product not found!'));
+      }
+
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: productId, userId: req.user._id });
+    })
     .then(() => {
       res.redirect('/admin/products');
     })
